refactor: rename `Header` to `SecurityHeader`
BREAKING CHANGE: `Header` has been renamed to `SecurityHeader`
