--- src/components/Header/migrate-to-2.x.md
@@ -1,3 +1,0 @@
-# `Header`
-
-`HeaderButton` has been deprecated - please use `Button` instead.
