+++ src/components/Panel/_mixins.scss
@@ -1,14 +1,14 @@
 ////
 /// Panel mixins.
 /// @group panel
-/// @copyright IBM Security 2019
+/// @copyright IBM Security 2020
 ////
 
 @import '../../globals/index';
 @import '../../globals/theme/mixins';
 
-@import '../../components/Header/mixins';
 @import '../../components/IconButton/mixins';
+@import '../../components/SecurityHeader/mixins';
 @import '../../components/Toolbar/mixins';
 
 @mixin panel {
