+++ src/components/SecurityHeader/HeaderListItem/_index.scss
