+++ src/components/SecurityHeader/HeaderNotification/_index.scss
