+++ src/components/SecurityHeader/HeaderNotification/_mixins.scss
