+++ src/components/SecurityHeader/HeaderPopoverHeader/_index.scss
