+++ src/components/SecurityHeader/HeaderPopoverHeader/_mixins.scss
