+++ src/components/SecurityHeader/HeaderPopoverLinkSecondary/_mixins.scss
