+++ src/components/SecurityHeader/_index.scss
