+++ src/components/SecurityHeader/_mixins.scss
