+++ src/components/SecurityHeader/index.js
@@ -1,7 +1,7 @@
 /**
  * @file Header entry point.
- * @copyright IBM Security 2019
+ * @copyright IBM Security 2020
  */
 
-export { default } from './Header';
+export { default } from './SecurityHeader';
 export { default as HeaderListItem } from './HeaderListItem';
