+++ src/components/SecurityHeader/locales/en/SecurityHeader.json
