+++ src/components/SecurityHeader/migrate-to-2.x.md
@@ -1,0 +1,7 @@
+# `Header`
+
+`Header` has been renamed to `SecurityHeader` to avoid conflicts with Carbon's `Header` component.
+
+## `HeaderButton`
+
+`HeaderButton` has been deprecated - please use `Button` instead.
