+++ src/components/Shell/_index.scss
@@ -1,12 +1,12 @@
 ////
 /// Shell component.
 /// @group shell
-/// @copyright IBM Security 2019
+/// @copyright IBM Security 2020
 ////
 
 @import '../Button/index';
-@import '../Header/index';
 @import '../Link/index';
+@import '../SecurityHeader/index';
 @import '../Toolbar/index';
 
 @import '../Component/mixins';
